Move Signup to routes directory
